# swagger/swagger.yaml
basePath: '{{.BasePath}}'
definitions:
...
  model.DeleteAsset:
    properties:
      domain:
        description: 数据域
        example: catalog
        type: string
      ownaddr:
        description: 资源的归属者
        example: ownaddr
        type: string
      reslist:
        description: 资源列表
        items:
          $ref: '#/definitions/model.ResList'
        type: array
    type: object
...
  model.ResList:
    properties:
      resids:
        example:
        - "123456"
        items:
          type: string
        type: array
      restype:
        example: data
        type: string
    type: object
...
  proto.Response:
    properties:
      code:
        description: 返回0成功 其中Data为其返回数据； 其他失败
        example: 0
        type: integer
      data:
        type: object
      message:
        description: 成功"OK"； 失败为错误信息
        example: OK
        type: string
      warn:
        type: string
    type: object
host: '{{.Host}}'
info:
  contact: {}
  description: '{{.Description}}'
  license: {}
  title: '{{.Title}}'
  version: '{{.Version}}'
paths:
...
  /api/v1/res/deleteResAuth:
    post:
      consumes:
      - application/json
      description: 资源归属者删除权限
      parameters:
      - description: token令牌
        in: header
        name: RAC_TOKEN
        type: string
      - description: 请求信息
        in: body
        name: deleteResAuth
        required: true
        schema:
          $ref: '#/definitions/model.DeleteAsset'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/proto.Response'
            type: object
      tags:
      - Auth
...
swagger: "2.0"